Introduce new prometheus metrics.
Add task queuing time which was previously part of the monitor endpoint (same calculation = task.started - task.received).
Add worker online gauge. We have the api/workers?status=true endpoint but we need something more useful in grafana that is scraped by prometheus and can show a nice graph.

Moved saving the event to the first line of the event method to have the freshest state available.
Removed unused imports.

Add handling of events required to produce the new metrics.

Add small tests to prove metrics work.
For queuing time it makes no sense to assert on the reported queuing time as it depends on environment...