Improvements to the prometheus integration documentation.
I have been asked for a video version of full integration from celery to flower to prometheus to grafana by members of the community.
This made me think that to get this all set up is not so trivial and better manual could help :). I don't do videos but like good manuals.

So here we go - step by step docs with lots of pics to make it easy.

I deleted all my set up and re-created it using these instructions and all worked so hopefully others can do same.