Add is_broker_connected function.
This function is called now before flower.start() is called to make sure we have a working broker and starting Flower will set it up correctly.
This is a first attempt at solving the issue of Flower booting up in a broken state.

There can be a race condition from the moment we checked the condition to the moment we call flower.start() although unlikely.
I will ponder on that...