Add number_of_tasks_queuing_at_worker metric.
This will show number of tasks fetched to the worker. Since Celery fetches prefetch_multiplier x concurrency setting number of tasks this will effectively cap out at that number.
This is useful to see if we hit the the cap and have to upgrade the hardware horizontally or vertically.

Fix last recorded task queuing time being used forever.

In case if a given worker recorded some queuing time for task with task name and then we never received any more tasks with that name on that worker the last recorded queuing time
would be used forever and the graph would never go back to zero.
Now if the task succeeds or fails it resets the queuing time.