Adding amqps as a valid broker scheme
Issue here: https://github.com/mher/flower/issues/1369

Flower does not take into account the `BROKER_API` config setting right
now if the scheme is set to `amqps`. It errors.

This seems like it should not be the case, as there is no change
required for the RabbitMQ API to work, regardless of amqp or amqps.